refactor: durationText抽离到util # Reviewed, transaction id: 23118
